Add Convolutional Code So Far
Here is the file containing the convolutional network code. Parts of our original code need to be integrated into the network in order to work, so I've appended the code from our previous network in while trying to figure out what we need to keep and what we need to get rid of. There is currently a modelling issue somewhere from lines 161 to 167, so the convolutional neural network is not yet operational.