Do Duplication Test on Resplit
I duplicated our resplit data, and then resplit that between training and testing. This manages to get the validation accuracy up to around 50%, but after 16 epochs, the training accuracy goes to 100%, effectively stopping any learning from happening. If we have time later, we might be able to run some regularization techniques on it to scale down the overfitting. However, I think this would help drive down the point that there should be more unique data added to the project.